+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 ## Dev/Test/CI stuff
 - Improved API tests runtime !3506 @alex.simm
 - Removed redirect after submit in foodsharepoint edit form to fix test !3527 @chriswalg
+- Port the register controller to Symfony !3594 @alex.simm
 
 # Hotfix 2024-08-29